engine/engine_impl.go: Use a patch request for image updates when starting steps.
When a pipeline fans out a DAG, concurrent calls are made to (*Kubernetes).Run
to start each step in parallel. Previously, starting a step required a
Read/Modify/Write loop to the Kubernetes API, and so updates were highly likely
to fail with a concurrent modification error.

This commit changes (*Kubernetes).start to use a PATCH request to update the
relevant container image of the Pod. The end result is that we are no longer at
risk of concurrent modification failures and can achieve higher performance on
getting pipeline stages started in a fan-out scenario.

A major downside of this approach is that "patch" is a separate RBAC permission
from "update". Existing installation instructions[1] do not include the
necessary permission, and so if this landed in master and users upgraded their
image, their existing runners may fail to work until they update the role
appropriately.

[1: https://docs.drone.io/runner/kubernetes/installation/]